# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an Import Export Backend API server for managing products, inventory, suppliers, and orders.",
        "title": "Import Export Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/inventories/submissions/{id}/close": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin/accountant closes an open reconciliation (open -\u003e closed). Staff can no longer edit child items; admin/accountant may still edit, then start processing or reopen.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "summary": "Close reconciliation submission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Submission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.InventorySubmission"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/inventories/submissions/{id}/reconciliation-items": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Staff submits counted quantities as a new in_progress child item under an initiated reconcile submission. Counts must not exceed the per-item snapshot baseline.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "summary": "Create reconciliation item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent submission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Counted items",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateReconciliationItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ReconciliationRequestItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/inventories/submissions/{id}/reconciliation-items/{item_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Staff replaces the counted quantities of their own child item. Editing a ready/approved row resets it to in_progress.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "summary": "Update reconciliation item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent submission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Child item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Counted items",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateReconciliationItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ReconciliationRequestItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Staff soft-deletes their own in_progress or ready child item. Approved/applied items cannot be deleted.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "summary": "Delete reconciliation item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent submission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Child item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/inventories/submissions/{id}/reopen": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin/accountant reopens a closed reconciliation (closed -\u003e open) so staff can edit child items again.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "summary": "Reopen reconciliation submission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Submission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.InventorySubmission"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/inventories/submissions/{id}/start-processing": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin/accountant applies a closed reconciliation in one atomic transaction. Re-checks for a consuming submission processed during the reconciliation window; on drift it rolls back and returns a warning payload (HTTP 409). Otherwise it creates the consuming transactions (snapshot - counted) and finalizes the submission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "summary": "Start processing a reconciliation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Submission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StartProcessingResult"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.StartProcessingResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateReconciliationItemRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.ReconciliationCountItem"
                    }
                }
            }
        },
...
        "dto.ReconciliationCountItem": {
            "type": "object",
            "required": [
                "inventory_item_id"
            ],
            "properties": {
                "inventory_item_id": {
                    "type": "integer"
                },
                "quantity": {
                    "description": "Quantity is a pointer so an omitted `quantity` (nil) is distinguishable from\nan explicit zero count. It deliberately carries NO validate:\"required\" tag:\nthe binding-layer validator would otherwise short-circuit a missing/null\nquantity into a generic validation error before the service runs, hiding the\nlocalized recon_item_missing_quantity domain error. Letting nil through to\nthe service's nil-check (validateCountsAgainstSnapshot) yields that localized\nerror instead. An explicit 0 still binds as a non-nil pointer and is a valid\ncount; the service validates the dereferenced value (non-negative, \u003c=\nsnapshot baseline).",
                    "type": "number"
                }
            }
        },
...
        "dto.StartProcessingResult": {
            "type": "object",
            "properties": {
                "drift_detected": {
                    "type": "boolean"
                },
                "submission": {
                    "$ref": "#/definitions/models.InventorySubmission"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.UpdateReconciliationItemRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.ReconciliationCountItem"
                    }
                }
            }
        },
...
        "models.Inventory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.InventoryItem"
                    }
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "restricted_delivery_hours": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.InventoryStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
        "models.InventoryItem": {
            "type": "object",
            "properties": {
                "active_purchase_transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.InventoryTransaction"
                    }
                },
                "consuming_transaction_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inventory": {
                    "$ref": "#/definitions/models.Inventory"
                },
                "inventory_id": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/models.Product"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/models.InventoryItemStatus"
                },
                "unit": {
                    "$ref": "#/definitions/models.Unit"
                },
                "unit_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "models.InventorySubmission": {
            "type": "object",
            "properties": {
                "approval_status": {
                    "$ref": "#/definitions/models.SubmissionApprovalStatus"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "error": {
                    "type": "object"
                },
                "id": {
                    "type": "integer"
                },
                "inventory": {
                    "$ref": "#/definitions/models.Inventory"
                },
                "inventory_id": {
                    "type": "integer"
                },
                "payload": {
                    "type": "object"
                },
                "processed_at": {
                    "description": "ProcessedAt is the precise instant a CONSUMING submission's processing\ncompleted (epic #38, Part 6, locked decision Q6). It is the authoritative\nwindow bound for the Start-Processing drift re-check (a sibling consuming\nsubmission with processed_at inside [snapshot_capture, now] is drift). Set\nwhen processing completes; nil while pending/failed.",
                    "type": "string"
                },
                "processing_status": {
                    "$ref": "#/definitions/models.SubmissionProcessingStatus"
                },
                "reason": {
                    "type": "string"
                },
                "reconcile_status": {
                    "description": "ReconcileStatus is the reconciliation lifecycle status (epic #38, Part 6).\nSet only for initiated reconciles (open at initiate); empty for every other\nsubmission type/flow. Drives the staff-immutability guard and the\nclose/reopen/start-processing transitions.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ReconcileLifecycleStatus"
                        }
                    ]
                },
                "submission_type": {
                    "$ref": "#/definitions/models.SubmissionType"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "models.ReconcileLifecycleStatus": {
            "type": "string",
            "enum": [
                "open",
                "closed",
                "processing",
                "processed"
            ],
            "x-enum-varnames": [
                "ReconcileLifecycleStatusOpen",
                "ReconcileLifecycleStatusClosed",
                "ReconcileLifecycleStatusProcessing",
                "ReconcileLifecycleStatusProcessed"
            ]
        },
        "models.ReconciliationRequestItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "payload": {
                    "type": "object"
                },
                "status": {
                    "$ref": "#/definitions/models.ReconciliationRequestItemStatus"
                },
                "submission": {
                    "$ref": "#/definitions/models.InventorySubmission"
                },
                "submission_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
        "models.ReconciliationRequestItemStatus": {
            "type": "string",
            "enum": [
                "in_progress"
            ],
            "x-enum-varnames": [
                "ReconciliationRequestItemStatusInProgress"
            ]
        },
...
        "models.SubmissionApprovalStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "InventorySubmissionApprovalStatusPending",
                "InventorySubmissionApprovalStatusApproved",
                "InventorySubmissionApprovalStatusRejected"
            ]
        },
        "models.SubmissionProcessingStatus": {
            "type": "string",
            "enum": [
                "pending",
                "failed",
                "completed",
                "canceled"
            ],
            "x-enum-varnames": [
                "InventorySubmissionStatusPending",
                "InventorySubmissionStatusFailed",
                "InventorySubmissionStatusCompleted",
                "InventorySubmissionStatusCanceled"
            ]
        },
        "models.SubmissionType": {
            "type": "string",
            "enum": [
                "reconcile",
                "dispose",
                "transfer"
            ],
            "x-enum-varnames": [
                "InventorySubmissionTypeReconcile",
                "InventorySubmissionTypeDispose",
                "InventorySubmissionTypeTransfer"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}